Demo 4 data processing: Trapezium area using float

+++ data_processing_02/src/Main.java
@@ -3,12 +3,12 @@ public class Main {
 
 	public static void main(String[] args) {
 
-		double b, B, h, area;
-		b = 6.0;
-		B = 8.0;
-		h = 5.0;
+		float b, B, h, area;
+		b = 6f;
+		B = 8f;
+		h = 5f;
 
-		area = (b + B) / 2.0 * h;
+		area = (b + B) / 2f * h;
 
 		System.out.println(area);// OUTPUT:35.0
 	}
